Simplify Block further
