+++ nemesys/modelling/blocks/block.py
@@ -1,12 +1,13 @@
-import copy
-from typing import Any, Iterable, Optional, Tuple, Union
-
-from nemesys.utils.conversion import ShapeConversion
+from typing import Any, Optional
 
 
 class Block:
     @property
-    def data(self) -> Any:
+    def data(self):
+        raise NotImplementedError
+
+    @data.setter
+    def data(self):
         raise NotImplementedError
 
     @staticmethod
@@ -15,20 +16,3 @@ class Block:
 
     def default(self):
         raise NotImplementedError
-
-    def read(self, key: Optional[Any]) -> Any:
-        raise NotImplementedError
-
-    def write(self, content: Optional[Any]):
-        raise NotImplementedError
-
-
-class ShapedBlock(Block):
-    def __init__(self, base_shape: Union[int, Tuple[int, ...]]):
-        super().__init__()
-
-        self._base_shape = ShapeConversion.to_tuple(base_shape)
-
-    @property
-    def base_shape(self) -> Tuple[int, ...]:
-        return self._base_shape
